components/benchmark/requirements.txt
flask
PyYAML
requests
matplotlib
numpy
pillow
opencv-python
